LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-duplicate-names-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.


FAIL policy - duplicate names assert_throws_js: function "_ => trustedTypes.createPolicy("default", {})" did not throw
